requirements.txt
appnope==0.1.0
asgiref==3.2.7
backcall==0.2.0
colorama==0.4.3
decorator==4.4.2
Django==3.0.6
django-ckeditor==5.9.0
django-debug-toolbar==2.2
django-js-asset==1.2.2
django-widget-tweaks==1.4.8
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.2.5
